requirements.txt
torch >= 2.1.0
timm >= 0.9.12
pytorch-lightning >= 2.1.3
torchvision >= 0.16.0
torchmetrics >= 1.2.1
numpy >= 1.26.0
opendatasets >= 0.1.22
hydra-core >= 1.3.2
pytest >= 7.4.4
dvc >= 3.39.0
dvc-gdrive >= 3.0.1